requirements.txt
dbt-postgres==1.7.10
watchdog==6.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
SQLAlchemy==2.0.38
sqlfluff==3.3.1
pytest-cov==6.0.0